Remove availability declaration from capabilities example
Remove the stanza declaring the VOSI-availability from the
VOSI-capabilities example.  The example wasn't wrong, since availability
may still optionally be present and declared, but the majority
of new services will probably not include availability,
and the example is more compact without it.
You could also do something similar in section 1.2.