Adding language on POST payloads containing parameters.
It was not easy picking a place for this.  If you add it where we
already talk about POST-s, it would appear twice (in sync and async).
If you add it where we talk about parameter encoding, it would awkwardly
sit next to VOTable content.

In the end, I think this location is the least bad one.

Content-wise, I considered whether to warn people that in urlencoded
parameters they have to replace + with space and in multipart parameters
they don't, but I figured that for almost everyone, this is done by
some library.  And for that reason, I also don't think we should think about
it much more -- *unless* this is becoming an issue and http server
implementations don't necessarily support urlencoded any more.  Do we
have indications that that is the case?